pyproject.toml
[project]
name = "investment-team"
version = "1.0.0"
requires-python = ">=3.12,<4"
readme = "README.md"
license-files = ["LICENSE"]
authors = [{ name = "Agno", email = "hello@agno.com" }]

dependencies = [
  "agno==2.5.0",
  "fastapi[standard]",
  "jiter==0.5.0",
  "mcp",
  "ollama",
  "openai==1.30.0",
  "pgvector",
  "psycopg[binary]",
  "python-dotenv",
  "sqlalchemy",
  "yfinance",
  # "parallel-web",  # Optional: uncomment if using ParallelTools
]

[project.optional-dependencies]
dev = ["mypy", "ruff"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agents*", "context*", "db*", "app*", "teams*", "workflows*"]

[tool.ruff]
line-length = 120
exclude = [".venv*"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = [".venv*"]

[[tool.mypy.overrides]]
module = ["pgvector.*", "setuptools.*", "agno.*", "httpx.*", "yfinance.*", "google.*"]
ignore_missing_imports = true

[tool.uv.pip]
no-annotate = true
